Code cleanup:         * Fix warnings in PdfCache         * Tiny cleanup in XournalWidget         * Move ElementContainer from view to model         * Const qualify getters

+++ src/core/control/PdfCache.h
@@ -26,7 +26,7 @@ class PdfCacheEntry;
 
 class PdfCache {
 public:
-    PdfCache(int size);
+    PdfCache(size_t size);
     virtual ~PdfCache();
 
 private:
